backend/poll-generation/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
arxiv==2.1.3
asttokens==3.0.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
exa-py==1.11.0
executing==2.2.0
fastapi==0.115.12
feedparser==6.0.11
frozenlist==1.5.0
groq==0.22.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-anthropic==0.3.10
langchain-community==0.3.21
langchain-core==0.3.51
langchain-groq==0.3.2
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langgraph==0.3.25
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.24
linkup-sdk==0.2.4
lxml==5.3.1
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.3.2
mypy-extensions==1.0.0
numpy==2.2.4
open_deep_research==0.0.10
openai==1.70.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
primp==0.14.0
prompt_toolkit==3.0.50
propcache==0.3.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
PyMuPDF==1.25.5
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sgmllib3k==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.1
tavily-python==0.5.4
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
duckduckgo_search==7.5.5